Search and replace
Replaced unsigned int timestep -> uint64_t timestep.
Add basic `Simulation` tests for 64-bit steps.

+++ hoomd/Analyzer.h
@@ -66,7 +66,7 @@ class PYBIND11_EXPORT Analyzer
         /*! Derived classes will implement this method to calculate their results
             \param timestep Current time step of the simulation
             */
-        virtual void analyze(unsigned int timestep){}
+        virtual void analyze(uint64_t timestep){}
 
         //! Sets the profiler for the analyzer to use
         void setProfiler(std::shared_ptr<Profiler> prof);
