+++ hoomd/CellListStencil.cc
@@ -49,7 +49,7 @@ CellListStencil::~CellListStencil()
     m_cl->getCellWidthChangeSignal().disconnect<CellListStencil, &CellListStencil::requestCompute>(this);
     }
 
-void CellListStencil::compute(unsigned int timestep)
+void CellListStencil::compute(uint64_t timestep)
     {
     // guard against unnecessary calls
     if (!shouldCompute(timestep)) return;
@@ -171,7 +171,7 @@ void CellListStencil::compute(unsigned int timestep)
         m_prof->pop();
     }
 
-bool CellListStencil::shouldCompute(unsigned int timestep)
+bool CellListStencil::shouldCompute(uint64_t timestep)
     {
     if (m_compute_stencil)
         {
