+++ hoomd/ConstForceCompute.h
@@ -59,7 +59,7 @@ class PYBIND11_EXPORT ConstForceCompute : public ForceCompute
         void rearrangeForces();
 
         //! Actually compute the forces
-        virtual void computeForces(unsigned int timestep);
+        virtual void computeForces(uint64_t timestep);
 
     private:
