+++ hoomd/GetarDumpWriter.cc
@@ -337,7 +337,7 @@ namespace getardump{
             m_archive->close();
         }
 
-    void GetarDumpWriter::analyze(unsigned int timestep)
+    void GetarDumpWriter::analyze(uint64_t timestep)
         {
         const unsigned int shiftedTimestep(timestep - m_offset);
         bool neededSnapshots[9] = {false, false, false, false, false,
@@ -436,7 +436,7 @@ namespace getardump{
             }
         }
 
-    void GetarDumpWriter::write(GTAR::BulkWriter &writer, const GetarDumpDescription &desc, unsigned int timestep)
+    void GetarDumpWriter::write(GTAR::BulkWriter &writer, const GetarDumpDescription &desc, uint64_t timestep)
         {
         if(!m_archive)
             return;
@@ -449,7 +449,7 @@ namespace getardump{
             writeUniform(writer, desc, timestep);
         }
 
-    void GetarDumpWriter::writeIndividual(GTAR::BulkWriter &writer, const GetarDumpDescription &desc, unsigned int timestep)
+    void GetarDumpWriter::writeIndividual(GTAR::BulkWriter &writer, const GetarDumpDescription &desc, uint64_t timestep)
         {
         if(desc.m_prop == AngularMomentum)
             {
@@ -720,7 +720,7 @@ namespace getardump{
             }
         }
 
-    void GetarDumpWriter::writeUniform(GTAR::BulkWriter &writer, const GetarDumpDescription &desc, unsigned int timestep)
+    void GetarDumpWriter::writeUniform(GTAR::BulkWriter &writer, const GetarDumpDescription &desc, uint64_t timestep)
         {
         if(desc.m_prop == Box)
             {
@@ -757,7 +757,7 @@ namespace getardump{
             }
         }
 
-    void GetarDumpWriter::writeText(GTAR::BulkWriter &writer, const GetarDumpDescription &desc, unsigned int timestep)
+    void GetarDumpWriter::writeText(GTAR::BulkWriter &writer, const GetarDumpDescription &desc, uint64_t timestep)
         {
         if(desc.m_prop == TypeNames)
             {
