+++ hoomd/Updater.h
@@ -66,7 +66,7 @@ class PYBIND11_EXPORT Updater
         /*! Derived classes will implement this method to perform their specific update
             \param timestep Current time step of the simulation
         */
-        virtual void update(unsigned int timestep)  {};
+        virtual void update(uint64_t timestep)  {};
 
         //! Sets the profiler for the compute to use
         virtual void setProfiler(std::shared_ptr<Profiler> prof);
@@ -103,7 +103,7 @@ class PYBIND11_EXPORT Updater
 
             See Logger for more information on what this is about.
         */
-        virtual Scalar getLogValue(const std::string& quantity, unsigned int timestep)
+        virtual Scalar getLogValue(const std::string& quantity, uint64_t timestep)
             {
             return Scalar(0.0);
             }
@@ -130,7 +130,7 @@ class PYBIND11_EXPORT Updater
 
             See LogMatrix for more information on what this is about.
         */
-        virtual pybind11::array getLogMatrix(const std::string& quantity, unsigned int timestep)
+        virtual pybind11::array getLogMatrix(const std::string& quantity, uint64_t timestep)
             {
             unsigned char tmp[] = {0};
             return pybind11::array(0,tmp);
